Reorganize code to respect scio structure
