+++ scio-extra/src/main/scala/com/spotify/scio/extra/voyager/syntax/AllSyntax.scala
@@ -1,0 +1,3 @@
+package com.spotify.scio.extra.voyager.syntax
+
+trait AllSyntax extends ScioContextSyntax with SCollectionSyntax
